Continuous collision detection with canvas walls
Set up initial simulation. 
Many bugs and bad practices, but it works for now.
Adding collision between particles in the future.
Issue: When gravity is applied, there are no countering forces. Because of this, the particles end up bouncing with a high frequency when their velocities get low. Implemented a temporary dampinFactor to counter this.